+++ packages/plugin-divider/src/viewer.js
@@ -1,1 +1,2 @@
 export { typeMapper as dividerTypeMapper } from './typeMapper';
+export { DIVIDER_TYPE } from './constants';