+++ packages/plugin-html/src/viewer.js
@@ -1,1 +1,2 @@
 export { typeMapper as htmlTypeMapper } from './typeMapper';
+export { HTML_TYPE } from './types';